requirements.txt
python-docx
PyYAML
langchain-core
langchain-groq
langchain-openai
langchain-google-genai
langchain-ollama